+++ app/src/main/java/com/example/expensediary/data/Expense.kt
@@ -20,6 +20,6 @@ data class Expense(
     @PrimaryKey(autoGenerate = true) var id: Long = 0L,
     @ColumnInfo(name = "user_id") var user_id: Long,
     @ColumnInfo(name = "item") var item: String,
-    @ColumnInfo(name = "price") var price: Int,
+    @ColumnInfo(name = "price") var price: Double,
     @ColumnInfo(name = "expense_date") var expenseDate: Date
 )
